refactor: standardize all amounts to decimal(15,2) and consolidate migrations
BREAKING CHANGE: Complete decimal refactor requires fresh migrations

## Currency Standardization
- Changed all amount columns from bigInteger to decimal(15,2)
- Removed all conversion logic (* 100 / 100) from entire codebase
- Updated all model casts to use decimal:2
- Fixed payment allocation display in frontend
- Cleaned up all 'cents'/'paise' references

## Migration Consolidation
- Consolidated 12 add_* migrations into main table migrations
- Deleted fragmented column addition migrations
- All tables now have complete column definitions from start

## Files Changed (40 total)
### Database (15 files)
- Modified 10 migration files to use decimal
- Deleted 11 add_* migration files
- Added audit columns to main migrations
- Added all features (type, transport, discount, etc) to main tables

### Models (9 files)
- Fixed PaymentAllocation cast (integer → decimal:2)
- Fixed Plan cast (integer → decimal:2)
- Added discount to InvoiceItem
- Updated all amount-related comments

### Controllers (10 files)
- PaymentController: Removed conversion logic
- CashbookController: Removed division by 100
- PlanController: Cleaned up commented code
- InvoiceController: Kept valid tax % calculations

### Frontend (6 files)
- InvoiceDetailView: Fixed allocation amount display
- invoice.ts: Updated type comments
- Other UI files: Authorization changes

### New Features
- Added OCR services (OcrService, LlmService, ReceiptScanningService)
- Added installation scripts (install_ocr_linux.sh, install_ocr_mac.sh)
- Added API policies (InvoicePolicy, ProductPolicy)
- Added Blameable trait for audit columns
- Added /api/expenses/scan endpoint

## Verification
✅ Zero bigInteger usage
✅ Zero cents/paise references
✅ All amounts use decimal(15,2)
✅ No conversion logic anywhere
✅ 100% decimal consistency

+++ backend/database/migrations/2025_12_23_111410_create_plans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->uuid('id')->primary();
             $table->string('name');
             $table->string('slug')->unique();
-            $table->integer('price')->default(0); // Direct Amount (e.g. INR)
+            $table->decimal('price', 15, 2)->default(0); // Plan price (Rupees)
             $table->string('interval')->default('monthly'); // monthly, yearly
             $table->text('description')->nullable();
             $table->string('status')->default('active');
